+++ db/migrate/20251217040453_remove_picked_up_from_settings.rb
@@ -1,5 +1,7 @@
 class RemovePickedUpFromSettings < ActiveRecord::Migration[7.1]
   def change
-    remove_column :settings, :customer_visible_picked_up, :boolean
+    if column_exists?(:settings, :customer_visible_picked_up)
+      remove_column :settings, :customer_visible_picked_up, :boolean
+    end
   end
 end
